+++ ts/navigation/WalletNavigator.ts
@@ -8,8 +8,8 @@ import BancomatDetailScreen from "../features/wallet/bancomat/screen/BancomatDet
 import AddDigitalMethodScreen from "../features/wallet/onboarding/AddDigitalMethodScreen";
 import WalletAddBancomatNavigator from "../features/wallet/onboarding/bancomat/navigation/navigator";
 import WALLET_ONBOARDING_BANCOMAT_ROUTES from "../features/wallet/onboarding/bancomat/navigation/routes";
-import PaymentMethodOnboardingBPayNavigator from "../features/wallet/onboarding/bancomatpay/navigation/navigator";
-import WALLET_ONBOARDING_BPAY_ROUTES from "../features/wallet/onboarding/bancomatpay/navigation/routes";
+import PaymentMethodOnboardingBPayNavigator from "../features/wallet/onboarding/bancomatPay/navigation/navigator";
+import WALLET_ONBOARDING_BPAY_ROUTES from "../features/wallet/onboarding/bancomatPay/navigation/routes";
 import PaymentMethodOnboardingSatispayNavigator from "../features/wallet/onboarding/satispay/navigation/navigator";
 import WALLET_ONBOARDING_SATISPAY_ROUTES from "../features/wallet/onboarding/satispay/navigation/routes";
 import SatispayDetailScreen from "../features/wallet/satispay/screen/SatispayDetailScreen";
