+++ ts/store/actions/types.ts
@@ -11,7 +11,7 @@ import {
 import { BonusActions } from "../../features/bonus/bonusVacanze/store/actions/bonusVacanze";
 import { BpdActions } from "../../features/bonus/bpd/store/actions";
 import { AbiActions } from "../../features/wallet/onboarding/bancomat/store/actions";
-import { BPayActions } from "../../features/wallet/onboarding/bancomatpay/store/actions";
+import { BPayActions } from "../../features/wallet/onboarding/bancomatPay/store/actions";
 import { SatispayActions } from "../../features/wallet/onboarding/satispay/store/actions";
 import { GlobalState } from "../reducers/types";
 import { AnalyticsActions } from "./analytics";
